[image-editor-sizing] Move dimensions input into new file
