+++ packages/perseus-editor/src/widgets/image-editor/components/image-settings.tsx
@@ -1,14 +1,12 @@
 import {components} from "@khanacademy/perseus";
 import {isFeatureOn} from "@khanacademy/perseus-core";
 import {LabeledField} from "@khanacademy/wonder-blocks-labeled-field";
-import {sizing} from "@khanacademy/wonder-blocks-tokens";
 import * as React from "react";
 
 import {AutoResizingTextArea} from "../../../components/auto-resizing-text-area";
-import ScrolllessNumberTextField from "../../../components/scrollless-number-text-field";
-import {getOtherSideLengthWithPreservedAspectRatio} from "../utils";
+import {wbFieldStyles} from "../utils";
 
-import styles from "./image-editor.module.css";
+import ImageDimensionsInput from "./image-dimensions-input";
 
 import type {Props} from "../image-editor";
 
@@ -33,46 +31,6 @@ export default function ImageSettings({
         return null;
     }
 
-    function handleWidthChange(newWidth: string) {
-        const newHeight = getOtherSideLengthWithPreservedAspectRatio(
-            backgroundImage.width!, // current side (width)
-            backgroundImage.height!, // other side (height)
-            Number(newWidth), // new side (width)
-        );
-
-        if (isNaN(newHeight)) {
-            return;
-        }
-
-        onChange({
-            backgroundImage: {
-                ...backgroundImage,
-                width: Number(newWidth),
-                height: newHeight,
-            },
-        });
-    }
-
-    function handleHeightChange(newHeight: string) {
-        const newWidth = getOtherSideLengthWithPreservedAspectRatio(
-            backgroundImage.height!, // current side (height)
-            backgroundImage.width!, // other side (width)
-            Number(newHeight), // new side (height)
-        );
-
-        if (isNaN(newWidth)) {
-            return;
-        }
-
-        onChange({
-            backgroundImage: {
-                ...backgroundImage,
-                height: Number(newHeight),
-                width: newWidth,
-            },
-        });
-    }
-
     return (
         <>
             {/* Preview */}
@@ -88,29 +46,10 @@ export default function ImageSettings({
             />
 
             {/* Dimensions */}
-            <div className={styles.dimensionsContainer}>
-                <LabeledField
-                    label="Width"
-                    field={
-                        <ScrolllessNumberTextField
-                            value={backgroundImage.width?.toString() ?? ""}
-                            onChange={handleWidthChange}
-                        />
-                    }
-                    styles={wbFieldStyles}
-                />
-                <span className={styles.xSpan}>x</span>
-                <LabeledField
-                    label="Height"
-                    field={
-                        <ScrolllessNumberTextField
-                            value={backgroundImage.height?.toString() ?? ""}
-                            onChange={handleHeightChange}
-                        />
-                    }
-                    styles={wbFieldStyles}
-                />
-            </div>
+            <ImageDimensionsInput
+                backgroundImage={backgroundImage}
+                onChange={onChange}
+            />
 
             {/* Alt text */}
             <LabeledField
@@ -166,14 +105,3 @@ export default function ImageSettings({
         </>
     );
 }
-
-// TODO: Use CSS modules after Wonder Blocks styles
-// are moved to a different layer.
-const wbFieldStyles = {
-    root: {
-        marginBlockEnd: sizing.size_080,
-    },
-    label: {
-        paddingBlockEnd: sizing.size_040,
-    },
-};
